logtail: perform zstd compression with single-segment mode
The "single-segment" mode encodes the expected uncompressed size
in the zstd frame, which the server can use as an optimization.

Also, add a new Config.CompressTransport flag.
Compression is an implementation detail of log uploading.
It could be zstd today or a different format in the future.
Most users of logtail should not care what compression is used.

Signed-off-by: Joe Tsai <joetsai@digital-static.net>