util/deephash: completely refactor
DO NOT SUBMIT

This is an experiment to see where we would want this to go.
If we're already using unsafe to hash memory,
then embrace the use of unsafe and stay out of reflect.

Rather than a typeHasherFunc that operates on

	func(*hasher, addressableValue)

we operate on

	func(*hasher, unsafe.Pointer)

where the unsafe.Pointer is identical to addressableValue.UnsafePointer.

We drop support for interface{ AppendTo([]byte) []byte }.
While that was more efficient in the prior pure-reflect implementation,
it is no longer true today. Instead, rely on direct memory hashing
of crypto-key types that have this method, and/or add direct
support for types that cannot be memory hashed
(for example time.Time and netip.Addr).

Performance:

	Hash-24                18.4µs ± 1%    13.5µs ± 1%  -26.82%  (p=0.000 n=9+9)
	HashPacketFilter-24    2.52µs ± 3%    2.24µs ± 1%  -11.14%  (p=0.000 n=10+10)
	TailcfgNode-24         1.42µs ± 1%    1.35µs ± 1%   -4.99%  (p=0.000 n=10+10)
	HashArray-24            324ns ± 1%     312ns ± 1%   -3.76%  (p=0.000 n=9+10)

SHA-256 hashing now takes up ~55% of the benchmark time.
We will get an additional boost in performance when Go's crypto/sha256
package uses native instructions. See https://go.dev/cl/408795.

Signed-off-by: Joe Tsai <joetsai@digital-static.net>