util/deephash: specialize hashing for netip.Addr and drop AppendTo support
There are 5 types that we care about that supports AppendTo:

	key.DiscoPublic
	key.NodePublic
	netip.Prefix
	netipx.IPRange
	netip.Addr

The key types are thin wrappers around [32]byte and are memory hashable.
The netip.Prefix and netipx.IPRange types are thin wrappers over netip.Addr.
The netip.Addr type is the only one with a complex structure where
the default behavior of deephash does not hash it correctly due to the presence
of the intern.Value type.

Drop support for AppendTo and instead add specialized hashing for netip.Addr
that would be semantically equivalent to == the netip.Addr values.

The AppendTo support was already broken prior to this change.
It was removed (intentionally or not) in #4870.
It was partially restored in #4858 for the fast path,
but still left out in the slow path.
Just drop support for it altogether.

This does mean we lack any ability for types to self-hash themselves.
In the future we can add support for types that implement:

	interface { DeepHash() Sum }

Test and fuzz cases were added for the relevant types that
used to rely on the AppendTo method.
FuzzAddr has been executed on 1 billion samples without issues.

Performance:

	name                 old time/op    new time/op    delta
	Hash-24                18.6µs ± 1%    18.2µs ± 1%  -1.66%  (p=0.000 n=9+10)
	HashPacketFilter-24    2.51µs ± 1%    2.50µs ± 1%    ~     (p=0.118 n=10+10)
	HashMapAcyclic-24      31.8µs ± 2%    31.0µs ± 1%  -2.36%  (p=0.000 n=10+9)
	TailcfgNode-24         1.43µs ± 1%    1.43µs ± 1%    ~     (p=0.753 n=10+10)
	HashArray-24            325ns ± 2%     324ns ± 0%    ~     (p=0.811 n=10+8)

Signed-off-by: Joe Tsai <joetsai@digital-static.net>