cmd/nginx-auth: ability to check application capabilities
A header `Expected-Cap` can now be passed which can be used to assert
that `WhoIsResponse.Caps` contains the provided capability.

This allows for ACL rules to define capabilities that can be checked
here in cmd/nginx-auth, meaning that ACLs can be used to control access
to various web services at Layer 7.

Signed-off-by: Jamie Greeff <jamie@greeff.me>