Add an option for tailscaled to disable the peerapi-dns service.
This is useful for exit node scenarios where we prefer DNS queries to go through
the regular exit node routing rather than handled by the DoH service.

Why would we ever prefer that? One reason is if we are already hosting our own
DNS infrastructure, and don't want to lose the association between clients and
the DNS queries they are performing for security/firewall/logging reasons.

Signed-off-by: Matrix User <matrixuser0@anon>